[METAED-1494] Drop MetaEd support below DS 3.2c (#189)
---------

+++ packages/metaed-plugin-edfi-ods-postgresql/test/integration/artifact/v7_1/Transcript-0020-Extension-Extension-Tables-Authoritative.sql
@@ -2,13 +2,13 @@
 CREATE TABLE extension.InstitutionControlDescriptor (
     InstitutionControlDescriptorId INT NOT NULL,
     CONSTRAINT InstitutionControlDescriptor_PK PRIMARY KEY (InstitutionControlDescriptorId)
-); 
+);
 
 -- Table extension.InstitutionLevelDescriptor --
 CREATE TABLE extension.InstitutionLevelDescriptor (
     InstitutionLevelDescriptorId INT NOT NULL,
     CONSTRAINT InstitutionLevelDescriptor_PK PRIMARY KEY (InstitutionLevelDescriptorId)
-); 
+);
 
 -- Table extension.PostSecondaryOrganization --
 CREATE TABLE extension.PostSecondaryOrganization (
@@ -21,7 +21,7 @@ CREATE TABLE extension.PostSecondaryOrganization (
     LastModifiedDate TIMESTAMP NOT NULL,
     Id UUID NOT NULL,
     CONSTRAINT PostSecondaryOrganization_PK PRIMARY KEY (NameOfInstitution)
-); 
+);
 ALTER TABLE extension.PostSecondaryOrganization ALTER COLUMN CreateDate SET DEFAULT current_timestamp;
 ALTER TABLE extension.PostSecondaryOrganization ALTER COLUMN Id SET DEFAULT gen_random_uuid();
 ALTER TABLE extension.PostSecondaryOrganization ALTER COLUMN LastModifiedDate SET DEFAULT current_timestamp;
@@ -30,23 +30,23 @@ ALTER TABLE extension.PostSecondaryOrganization ALTER COLUMN LastModifiedDate SE
 CREATE TABLE extension.SpecialEducationGraduationStatusDescriptor (
     SpecialEducationGraduationStatusDescriptorId INT NOT NULL,
     CONSTRAINT SpecialEducationGraduationStatusDescriptor_PK PRIMARY KEY (SpecialEducationGraduationStatusDescriptorId)
-); 
+);
 
 -- Table extension.StudentAcademicRecordClassRankingExtension --
 CREATE TABLE extension.StudentAcademicRecordClassRankingExtension (
-    EducationOrganizationId INT NOT NULL,
+    EducationOrganizationId BIGINT NOT NULL,
     SchoolYear SMALLINT NOT NULL,
     StudentUSI INT NOT NULL,
     TermDescriptorId INT NOT NULL,
     SpecialEducationGraduationStatusDescriptorId INT NOT NULL,
     CreateDate TIMESTAMP NOT NULL,
     CONSTRAINT StudentAcademicRecordClassRankingExtension_PK PRIMARY KEY (EducationOrganizationId, SchoolYear, StudentUSI, TermDescriptorId)
-); 
+);
 ALTER TABLE extension.StudentAcademicRecordClassRankingExtension ALTER COLUMN CreateDate SET DEFAULT current_timestamp;
 
 -- Table extension.StudentAcademicRecordExtension --
 CREATE TABLE extension.StudentAcademicRecordExtension (
-    EducationOrganizationId INT NOT NULL,
+    EducationOrganizationId BIGINT NOT NULL,
     SchoolYear SMALLINT NOT NULL,
     StudentUSI INT NOT NULL,
     TermDescriptorId INT NOT NULL,
@@ -54,12 +54,12 @@ CREATE TABLE extension.StudentAcademicRecordExtension (
     SubmissionCertificationDescriptorId INT NULL,
     CreateDate TIMESTAMP NOT NULL,
     CONSTRAINT StudentAcademicRecordExtension_PK PRIMARY KEY (EducationOrganizationId, SchoolYear, StudentUSI, TermDescriptorId)
-); 
+);
 ALTER TABLE extension.StudentAcademicRecordExtension ALTER COLUMN CreateDate SET DEFAULT current_timestamp;
 
 -- Table extension.SubmissionCertificationDescriptor --
 CREATE TABLE extension.SubmissionCertificationDescriptor (
     SubmissionCertificationDescriptorId INT NOT NULL,
     CONSTRAINT SubmissionCertificationDescriptor_PK PRIMARY KEY (SubmissionCertificationDescriptorId)
-); 
+);
